bug 1401246: Standardize signal handlers and apps
Move signal handlers into signal_handlers.py, standardize imports and
usage, and use the @receiver decorator to register them. Use
AppConfig.ready to import and register signal handlers, and use the
dotted paths to AppConfigs in INSTALLED_APPS.
