+++ kuma/search/decorators.py
@@ -1,12 +1,6 @@
 import logging
 
-from django.db.models.signals import pre_delete
-
 from elasticsearch.exceptions import ConnectionError
-
-from kuma.wiki.signals import render_done
-
-from .signals import render_done_handler, pre_delete_handler
 
 
 log = logging.getLogger('kuma.search.decorators')
@@ -28,12 +22,3 @@ def requires_good_connection(fun):
                       'ready to rumble, is not running at all, or ES_URLS'
                       'is set wrong in your .env file.')
     return _requires_good_connection
-
-
-def register_live_index(model_cls):
-    """Register a model and index for auto indexing."""
-    uid = str(model_cls) + 'live_indexing'
-    render_done.connect(render_done_handler, model_cls, dispatch_uid=uid)
-    pre_delete.connect(pre_delete_handler, model_cls, dispatch_uid=uid)
-    # Enable this to be used as decorator.
-    return model_cls
