+++ kuma/settings/common.py
@@ -573,10 +573,10 @@ INSTALLED_APPS = (
     'kuma.landing',
     'kuma.redirects',
     'kuma.scrape',
-    'kuma.search',
-    'kuma.users',
-    'kuma.wiki',
-    'kuma.attachments',
+    'kuma.search.apps.SearchConfig',
+    'kuma.users.apps.UserConfig',
+    'kuma.wiki.apps.WikiConfig',
+    'kuma.attachments.apps.AttachmentsConfig',
     'allauth',
     'allauth.account',
     'allauth.socialaccount',
