+++ kuma/users/__init__.py
@@ -1,2 +1,1 @@
-# work around for bug in Jingo: https://github.com/jbalogh/jingo/issues/68
-default_app_config = 'kuma.users.apps.UserConfig'
+"""kuma.users handles users, login, and profile data."""
