+++ kuma/wiki/__init__.py
@@ -1,2 +1,1 @@
-# work around for bug in Jingo: https://github.com/jbalogh/jingo/issues/68
-default_app_config = 'kuma.wiki.apps.WikiConfig'
+"""kuma.wiki handles wiki pages and related data."""
